// src/types/utils-data.ts
export type SponsorType = {
  id: number;
  name: string;
  image: string;
  type: string;
};

export type AboutType = {
  id: number;
  image: string;
  type: string;
};

export type FacilityType = {
  id: number;
  name: string;
};

export type TestimonialType = {
  id: number;
  username: string;
  title: string;
  image: string;
  testimoni: string;
  role: string;
  rating: number;
  createdAt: string;
};